Add dry run option

+++ disable_inactive.sh
@@ -3,6 +3,7 @@
 ###
 ### Set inputs
 ###
+dry_run="${{ inputs.dry-run }}"
 max_days_inactive="${{ inputs.max-days-inactive }}"
 gh_org="${{ inputs.github-org }}"
 GITHUB_TOKEN="${{ inputs.github-pat }}"
@@ -54,6 +55,10 @@ hold_until_rate_limit_success() {
 
 # Remove user from copilot
 remove_user_from_copilot() {
+  if [[ $dry_run == "true" ]]; then
+    echo "🧪 Would remove $copilot_user from CoPilot"
+    return 0
+  fi
   
   REMOVE_USER_FROM_COPILOT=$(curl -sL \
     -X DELETE \
@@ -77,6 +82,10 @@ remove_user_from_copilot() {
 ###
 ### Fetch users to iterate over
 ###
+
+if [[ $dry_run == "true" ]]; then
+  echo "🧪 Dry run is enabled.  Users won't be removed"
+fi
 
 # Get count of all seats
 copilot_seats_total_count=$(curl -s \
@@ -157,7 +166,7 @@ while IFS=$'\n' read -r copilot_user; do
   last_editor=$(echo "$user_data" | jq -r '.last_activity_editor')
   echo "Last editor: $last_editor"
 
-   # Get the last active date of the user
+  # Get the last active date of the user
   last_active_date=$(echo "$user_data" | jq -r '.last_activity_at')
   echo "Last activity date at: $last_active_date"
